made requested changes, updated unit tests
