+++ packages/peregrine/lib/talons/CartPage/ProductListing/useQuantity.js
@@ -1,9 +1,7 @@
-import { useCallback, useMemo, useState, useEffect } from 'react';
-import { useFieldApi } from 'informed';
-import useFieldState from '@magento/peregrine/lib/hooks/hook-wrappers/useInformedFieldStateWrapper';
-import debounce from 'lodash.debounce';
-
 /**
+ *
+ * @deprecated - use talons/QuantityStepper/useQuantityStepper instead
+ *
  * This talon contains logic for a product quantity UI component.
  * It performs effects and returns prop data for rendering a component that lets you
  * modify the quantity of a cart item.
@@ -24,91 +22,14 @@ import debounce from 'lodash.debounce';
  * @example <caption>Importing into your project</caption>
  * import { useQuantity } from '@magento/peregrine/lib/talons/CartPage/ProductListing/useQuantity';
  */
-export const useQuantity = props => {
-    const { initialValue, min, onChange } = props;
-
-    const [prevQuantity, setPrevQuantity] = useState(initialValue);
-
-    const quantityFieldApi = useFieldApi('quantity');
-    const { value: quantity } = useFieldState('quantity');
-
-    const isIncrementDisabled = useMemo(() => !quantity, [quantity]);
-
-    // "min: 0" lets a user delete the value and enter a new one, but "1" is
-    // actually the minimum value we allow to be set through decrement button.
-    const isDecrementDisabled = useMemo(() => !quantity || quantity <= 1, [
-        quantity
-    ]);
-
-    // Fire the onChange after some wait time. We calculate the current delay
-    // as enough time for a user to spam inc/dec quantity but not enough time
-    // for a user to click inc/dec on Product A and then click Product B.
-    const debouncedOnChange = useMemo(
-        () =>
-            debounce(val => {
-                setPrevQuantity(val);
-                onChange(val);
-            }, 350),
-        [onChange]
-    );
-
-    const handleDecrement = useCallback(() => {
-        const newQuantity = quantity - 1;
-        quantityFieldApi.setValue(newQuantity);
-        debouncedOnChange(newQuantity);
-    }, [debouncedOnChange, quantity, quantityFieldApi]);
-
-    const handleIncrement = useCallback(() => {
-        const newQuantity = quantity + 1;
-        quantityFieldApi.setValue(newQuantity);
-        debouncedOnChange(newQuantity);
-    }, [debouncedOnChange, quantity, quantityFieldApi]);
-
-    const handleBlur = useCallback(() => {
-        // Only submit the value change if it has changed.
-        if (typeof quantity === 'number' && quantity != prevQuantity) {
-            debouncedOnChange(quantity);
-        }
-    }, [debouncedOnChange, prevQuantity, quantity]);
-
-    const maskInput = useCallback(
-        value => {
-            try {
-                // For some storefronts decimal values are allowed.
-                const nextVal = parseFloat(value);
-                if (value && isNaN(nextVal))
-                    throw new Error(`${value} is not a number.`);
-                if (nextVal < min) return min;
-                else return nextVal;
-            } catch (err) {
-                console.error(err);
-                return prevQuantity;
-            }
-        },
-        [min, prevQuantity]
-    );
-
-    /**
-     * Everytime initialValue changes, update the quantity field state.
-     */
-    useEffect(() => {
-        quantityFieldApi.setValue(initialValue);
-    }, [initialValue, quantityFieldApi]);
-
-    return {
-        isDecrementDisabled,
-        isIncrementDisabled,
-        handleBlur,
-        handleDecrement,
-        handleIncrement,
-        maskInput
-    };
-};
+export {
+    useQuantityStepper as useQuantity
+} from '../../QuantityStepper/useQuantityStepper';
 
 /** JSDoc type definitions */
 
 /**
- * Object type returned by the {@link useQuantity} talon.
+ * Object type returned by the {@link useQuantityStepper} talon.
  * It provides props data for a quantity UI component.
  *
  * @typedef {Object} QuantityTalonProps
