+++ packages/venia-ui/lib/components/CartPage/ProductListing/quantity.js
@@ -1,94 +1,7 @@
-import React, { Fragment } from 'react';
-import { useIntl } from 'react-intl';
+import React from 'react';
 import { Form } from 'informed';
 import { func, number, string } from 'prop-types';
-import { Minus as MinusIcon, Plus as PlusIcon } from 'react-feather';
-import { useQuantity } from '@magento/peregrine/lib/talons/CartPage/ProductListing/useQuantity';
-
-import { useStyle } from '../../../classify';
-import Icon from '../../Icon';
-import TextInput from '../../TextInput';
-import { Message } from '../../Field';
-import defaultClasses from './quantity.module.css';
-
-/**
- * @deprecated - use components/QuantityStepper instead
- */
-export const QuantityFields = props => {
-    const { initialValue, itemId, label, min, onChange, message } = props;
-    const { formatMessage } = useIntl();
-    const classes = useStyle(defaultClasses, props.classes);
-    const iconClasses = { root: classes.icon };
-
-    const talonProps = useQuantity({
-        initialValue,
-        min,
-        onChange
-    });
-
-    const {
-        isDecrementDisabled,
-        isIncrementDisabled,
-        handleBlur,
-        handleDecrement,
-        handleIncrement,
-        maskInput
-    } = talonProps;
-
-    const errorMessage = message ? <Message>{message}</Message> : null;
-
-    return (
-        <Fragment>
-            <div className={classes.root}>
-                <label className={classes.label} htmlFor={itemId}>
-                    {label}
-                </label>
-                <button
-                    aria-label={formatMessage({
-                        id: 'quantity.buttonDecrement',
-                        defaultMessage: 'Decrease Quantity'
-                    })}
-                    className={classes.button_decrement}
-                    disabled={isDecrementDisabled}
-                    onClick={handleDecrement}
-                    type="button"
-                    data-cy="Quantity-decrementButton"
-                >
-                    <Icon classes={iconClasses} src={MinusIcon} size={22} />
-                </button>
-                <TextInput
-                    aria-label={formatMessage({
-                        id: 'quantity.input',
-                        defaultMessage: 'Item Quantity'
-                    })}
-                    data-cy="QuantityFields-input"
-                    classes={{ input: classes.input }}
-                    field="quantity"
-                    id={itemId}
-                    inputMode="numeric"
-                    mask={maskInput}
-                    min={min}
-                    onBlur={handleBlur}
-                    pattern="[0-9]*"
-                />
-                <button
-                    aria-label={formatMessage({
-                        id: 'quantity.buttonIncrement',
-                        defaultMessage: 'Increase Quantity'
-                    })}
-                    className={classes.button_increment}
-                    disabled={isIncrementDisabled}
-                    onClick={handleIncrement}
-                    type="button"
-                    data-cy="Quantity-incrementButton"
-                >
-                    <Icon classes={iconClasses} src={PlusIcon} size={20} />
-                </button>
-            </div>
-            {errorMessage}
-        </Fragment>
-    );
-};
+import QuantityStepper from '../../QuantityStepper';
 
 const Quantity = props => {
     return (
@@ -97,7 +10,7 @@ const Quantity = props => {
                 quantity: props.initialValue
             }}
         >
-            <QuantityFields {...props} />
+            <QuantityStepper {...props} />
         </Form>
     );
 };
@@ -118,10 +31,6 @@ Quantity.defaultProps = {
     onChange: () => {}
 };
 
-QuantityFields.defaultProps = {
-    min: 0,
-    initialValue: 1,
-    onChange: () => {}
-};
+export const QuantityFields = QuantityStepper;
 
 export default Quantity;
