--- packages/venia-ui/lib/components/CartPage/ProductListing/quantity.module.css
@@ -1,76 +1,0 @@
-.root {
-    --stroke: var(--venia-global-color-gray-600);
-    text-align: center;
-    display: grid;
-    grid-template-columns: 1fr 4rem 1fr;
-    grid-column-gap: 0.5rem;
-    align-items: center;
-    justify-items: center;
-}
-
-.wrap {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-}
-
-.label:not(:focus):not(:active) {
-    position: absolute;
-    overflow: hidden;
-    width: 1px;
-    height: 1px;
-    clip: rect(0 0 0 0);
-    clip-path: inset(100%);
-    white-space: nowrap;
-}
-
-.button {
-    align-items: center;
-    background-color: rgb(var(--venia-global-color-gray-50));
-    border: 2px solid rgb(var(--venia-global-color-gray-600));
-    border-radius: 50%;
-    display: inline-flex;
-    height: 2rem;
-    justify-content: center;
-    stroke: rgb(var(--stroke));
-    width: 2rem;
-}
-
-.button:hover {
-    --stroke: rgb(var(--venia-global-color-gray-900));
-    border-color: rgb(var(--stroke));
-}
-
-.button:disabled,
-.button:disabled:hover {
-    --stroke: rgb(var(--venia-global-color-gray-400));
-    border-color: rgb(var(--venia-global-color-gray-400));
-}
-
-.button:disabled {
-    cursor: not-allowed;
-}
-
-.button:focus {
-    outline: none;
-}
-
-.button_decrement {
-    composes: button;
-}
-
-.button_increment {
-    composes: button;
-}
-
-.icon {
-    align-items: center;
-    display: inline-flex;
-    justify-content: center;
-    touch-action: manipulation;
-}
-
-.input {
-    composes: input from '../../TextInput/textInput.module.css';
-    text-align: center;
-}
