+++ packages/venia-ui/lib/components/ProductFullDetail/productFullDetail.module.css
@@ -169,7 +169,7 @@
 }
 
 .quantityRoot {
-    composes: root from '../CartPage/ProductListing/quantity.module.css';
+    composes: root from '../QuantityStepper/quantityStepper.module.css';
     grid-template-columns: auto 4rem auto;
     justify-content: start;
 }
