+++ packages/venia-ui/lib/components/QuantityStepper/index.js
@@ -1,1 +1,1 @@
-export { default } from './QuantityStepper';
+export { default } from './quantityStepper';
